pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-exercises-generator"
version = "0.1.0"
description = "Explore ways to automatically generate programming exercises using AI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.14.0",
    "rich>=14.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
finetune = [
    "bitsandbytes>=0.49.0",
    "modal>=1.3.0.post1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "trl>=0.24.0",
    "unsloth>=2025.12.9",
    "wandb>=0.23.1",
]
modal = [
    "modal>=1.3.0.post1",
]

[project.scripts]
python-exercises-generator = "python_exercises_generator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ipython>=9.8.0",
]
